html/semantics/disabled-elements/disabled-elements-01.html

<!DOCTYPE html>
<title>Disabled Elements Test</title>
<link rel="author" title="Eric Casler" href="mailto:ericorange@gmail.com">
<link rel="help" href="http://www.whatwg.org/html5/#disabled-elements">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>
<div id="root" />
<script>
// Elements tested for in this file
var types = ["button", "input", "select", "textarea"];
// no tests for: optgroup, option, fieldset

var root = document.getElementById("root");
for(var element_type = 0; element_type < types.length; element_type++) {
	test(function() {
		root.innerHTML = "<"+types[element_type]+" + id='elem'/>";

		var elem = document.getElementById("elem");
		assert_false(elem.disabled);
	},
	"Test ["+types[element_type]+"]: default behaviour is NOT disabled");

	test(function() {
		var formats = ["disabled", "disabled=disabled",
			"disabled='disabled'", "disabled='true'",
			"disabled=true"];

		for(var f = 0; f < formats.length; f++) {
			root.innerHTML = "<"+types[element_type]+" id='elem' " + formats[f] + "/>";

			var elem = document.getElementById("elem");
			assert_true(elem.disabled);
		}
	},
	"Test ["+types[element_type]+"]: verify disabled acts as boolean attribute");

	test(function() {
		root.innerHTML = "<"+types[element_type]+" id='elem'></"+types[element_type]+"><input type='hidden' id='check' value='pass' />";
		var elem = document.getElementById("elem");

		assert_equals(check.value, "pass");

		var disabled_status = elem.disabled;
		elem.disabled = true;
		assert_true(elem.disabled);

		elem.onclick = function () {
			var check = document.getElementById("check");
			check.value = "fail";
		};

		var evObj = document.createEvent('Events');
		evObj.initEvent("click", true, false);
		elem.dispatchEvent(evObj);
		assert_equals(check.value, "pass");
	},
	"Test ["+types[element_type]+"]: click event dispatch should be prevented");
}
root.innerHTML = "";
</script>
